Add provider sync job
Improve tests for Provider import rake task

Lint fix
