+++ app/jobs/publish_teacher_training/provider/sync_all_job.rb
@@ -1,0 +1,11 @@
+module PublishTeacherTraining
+  module Provider
+    class SyncAllJob < ApplicationJob
+      queue_as :default
+
+      def perform
+        PublishTeacherTraining::Provider::Importer.call
+      end
+    end
+  end
+end
