+++ lib/tasks/provider_data.rake
@@ -1,6 +1,6 @@
 namespace :provider_data do
   desc "Import all Providers from Publish Teacher Training"
   task import: :environment do
-    Provider::Importer.call
+    PublishTeacherTraining::Provider::Importer.call
   end
 end
